allsocs: udma: Handling dual versions of udma.
- With the introdution of lcdma with AM62lx device family,
the behavior of udma is different from other k3 class of devices.
As a result, there exists now two versions for udma driver in
mcu-plus-sdk. In order to handle those changes, this commit,
handles the limitation of directly using udma driver headers and
fixes build issues.
- This commit also updates some of the makefile updates that are not
committed when performed gen-buildfiles.

Fixes: SITSW-7106

Signed-off-by: Soumya Tripathy <s-tripathy@ti.com>